requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-filter==2.0.0
djangorestframework==3.9.0
djangorestframework-csv==2.1.0
gunicorn==19.9.0
Markdown==3.0.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.7
static3==0.7.0
six==1.11.0
unicodecsv==0.14.1
whitenoise==4.1.2
